Added art and animations
Added the art and a bit of animation.
But for the slash I had to make a canvas (to contain the image) but i dont know how it would work now with placing it   on the right position. Also I hid the canvas because when press play it played once. I dont know why. I have a sprite with weapon_1, weapon_2.... enemy_1, enemy_2... I think this makes it easier in code to replace the current sprite with +1 or smth like that. But I only added tier 1 (weapon_1) weapon in the "inventory". The hearts are normally alpha 255. Maybe when losing one heart lower the opacity.